refactor(web): break Upload in multiple files
