+++ packages/web/src/components/navigation/upload/Upload.tsx
@@ -16,18 +16,16 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import React, {useRef} from 'react';
-
+import React from 'react';
 import {FiUpload} from 'react-icons/fi';
+
+import {useFormik} from 'formik';
 
 import {
   Button,
   chakra,
-  FormControl,
-  FormErrorMessage,
-  FormLabel,
+  HStack,
   IconButton,
-  Input,
   Modal,
   ModalBody,
   ModalCloseButton,
@@ -35,22 +33,15 @@ import {
   ModalFooter,
   ModalHeader,
   ModalOverlay,
-  VStack,
-  HStack,
-  Flex,
-  Heading,
   useDisclosure,
 } from '@chakra-ui/react';
 
-import {useFormik} from 'formik';
-
 import * as Yup from 'yup';
 
-type UploadForm = {
-  readonly title: string;
-  readonly description: string;
-  readonly video: File | null;
-}
+import VideoInputs from './VideoInputs';
+import DataInputs from './DataInputs';
+
+import UploadForm from './UploadForm';
 
 function Upload() {
   const {isOpen, onOpen, onClose} = useDisclosure();
@@ -102,93 +93,6 @@ function Upload() {
         </ModalContent>
       </Modal>
     </>
-  );
-}
-
-type DataInputsProps = {
-  readonly formik: ReturnType<typeof useFormik<UploadForm>>;
-};
-
-function VideoInputs(props: DataInputsProps) {
-  const {formik} = props;
-
-  const inputRef = useRef<HTMLInputElement>();
-
-  return (
-    <Flex
-      padding='0 1.5rem'
-      height='100%'
-      border='2px dashed #cecece'
-      borderRadius='1rem'
-      align='center'
-      justify='center'
-    >
-      <VStack>
-        <Heading as='h3' fontSize='1.2rem'>
-          Select video to upload
-        </Heading>
-
-        <chakra.span fontSize='0.85rem' color='#5c5c5c'>
-          Only MP4
-        </chakra.span>
-
-        <Button onClick={() => inputRef.current.click()}>
-          Select file
-        </Button>
-
-        <Input
-          hidden
-          ref={inputRef}
-          type='file'
-          name='video'
-          onChange={formik.handleChange}
-        />
-      </VStack>
-    </Flex>
-  );
-}
-
-function DataInputs(props: DataInputsProps) {
-  const {formik} = props;
-
-  return (
-    <VStack gap='0.5rem'>
-      <FormControl isInvalid={Boolean(formik.errors.title)}>
-        <FormLabel>Title</FormLabel>
-
-        <Input
-          name='title'
-          title='Title'
-          placeholder='Title'
-          value={formik.values.title}
-          onChange={formik.handleChange}
-        />
-
-        {formik.errors.title && (
-          <FormErrorMessage>
-            {formik.errors.title}
-          </FormErrorMessage>
-        )}
-      </FormControl>
-
-      <FormControl isInvalid={Boolean(formik.errors.description)}>
-        <FormLabel>Description</FormLabel>
-
-        <Input
-          name='description'
-          title='Description'
-          placeholder='Description'
-          value={formik.values.description}
-          onChange={formik.handleChange}
-        />
-
-        {formik.errors.description && (
-          <FormErrorMessage>
-            {formik.errors.description}
-          </FormErrorMessage>
-        )}
-      </FormControl>
-    </VStack>
   );
 }
